Fix standalone scaffolds by moving embedding into core
Standalone scaffolds of templates that embed apps (content, design,
assets) were uninstallable: the create --standalone path rewrites
workspace:* deps to latest, but @agent-native/embedding is intentionally
workspace-only and not published to npm, so pnpm install 404'd.

Move the embed surface (EmbeddedApp, bridge, agent, protocol) into the
published @agent-native/core under ./embedding subpath exports, point the
three templates at @agent-native/core/embedding/*, and drop their
@agent-native/embedding dependency. The @agent-native/embedding package
stays as a thin re-export for backward compatibility.

Fixes #1337
