+++ packages/core/package.json
@@ -116,6 +116,11 @@
     "./router": "./dist/router/index.js",
     "./collab": "./dist/collab/index.js",
     "./a2a": "./dist/a2a/index.js",
+    "./embedding": "./dist/embedding/index.js",
+    "./embedding/react": "./dist/embedding/react.js",
+    "./embedding/bridge": "./dist/embedding/bridge.js",
+    "./embedding/agent": "./dist/embedding/agent.js",
+    "./embedding/protocol": "./dist/embedding/protocol.js",
     "./mcp": "./dist/mcp/index.js",
     "./mcp-client": "./dist/mcp-client/index.js",
     "./tracking": "./dist/tracking/index.js",
