+++ packages/core/src/embedding/index.ts
@@ -1,0 +1,43 @@
+export {
+  EmbeddedApp,
+  type EmbeddedAppMessageInfo,
+  type EmbeddedAppProps,
+  type EmbeddedAppRef,
+} from "./react.js";
+export {
+  announceEmbeddedAppReady,
+  createEmbeddedAppBridge,
+  sendEmbeddedAppMessage,
+  type EmbeddedAppBridge,
+  type EmbeddedAppBridgeOptions,
+  type EmbeddedAppMessageEvent,
+  type EmbeddedAppMessageHandler,
+  type EmbeddedAppRequestHandler,
+} from "./bridge.js";
+export {
+  A2AClient,
+  getA2AUrl,
+  getAgentCardUrl,
+  getMcpUrl,
+  sendMessage,
+  type AgentEndpointOptions,
+  type Message,
+  type SendMessageOptions,
+  type Task,
+} from "./agent.js";
+export {
+  AGENT_NATIVE_EMBED_MESSAGE_TYPES,
+  AGENT_NATIVE_EMBED_PROTOCOL,
+  AGENT_NATIVE_EMBED_VERSION,
+  createAgentNativeEmbedEnvelope,
+  createEmbeddedAppRequestId,
+  embeddedAppOrigin,
+  isAgentNativeEmbedEnvelope,
+  isAllowedEmbeddedAppOrigin,
+  messageErrorPayload,
+  withEmbeddedAppParams,
+  type AgentNativeEmbedEnvelope,
+  type AgentNativeEmbedErrorPayload,
+  type AgentNativeEmbedMessageType,
+  type EmbeddedAppUrlOptions,
+} from "./protocol.js";
