+++ templates/content/app/components/editor/extensions/ImageBlock.tsx
@@ -10,7 +10,7 @@ import {
 import {
   EmbeddedApp,
   type EmbeddedAppRef,
-} from "@agent-native/embedding/react";
+} from "@agent-native/core/embedding/react";
 import {
   IconArrowsMaximize,
   IconArrowsMinimize,
