+++ templates/content/package.json
@@ -23,7 +23,6 @@
   },
   "dependencies": {
     "@agent-native/core": "workspace:*",
-    "@agent-native/embedding": "workspace:*",
     "@dnd-kit/core": "^6.3.1",
     "@dnd-kit/sortable": "^10.0.0",
     "@dnd-kit/utilities": "^3.2.2",
