+++ templates/design/package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "@agent-native/core": "workspace:*",
-    "@agent-native/embedding": "workspace:*",
     "@agent-native/pinpoint": "workspace:*",
     "@fontsource-variable/inter": "^5.2.8",
     "@libsql/client": "^0.15.8",
